% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_answer_shares.R
\name{inv_answer_shares}
\alias{inv_answer_shares}
\title{Test answer distribution}
\usage{
inv_answer_shares(
  survey_data,
  answer_options_to_exclude,
  alpha = 0.05,
  max_answer_options_of_question = 15,
  translate_table = NULL,
  output = "conspi"
)
}
\arguments{
\item{survey_data}{data table of class \strong{be_monk}}

\item{answer_options_to_exclude}{a vector of answer options that shall not be included in the analysis}

\item{alpha}{singnificance level on that shall be tested}

\item{max_answer_options_of_question}{threshold of number of answer options that the question shall be analysed}
}
\value{
data table with unadjusted and adjusted p-values for each question-answer-interviewer-combination
}
\description{
This function investigates the answer shares by interviewer and detects deviances in shares.
}
\details{
Each answer option is treated as a binary variable.
To test whether the share of a specific answer option for a specific interviewer deviates from the expected probability,  estimated from all interviews, a binomial test is carried out. p-values are p-adjusted. Significances are calculated on #'level alpha.
}
