% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_speeders_median.r
\name{inv_speeders_median}
\alias{inv_speeders_median}
\title{Test for fast Quantiles in Answer Times}
\usage{
inv_speeders_median(data, method = "q20", answer_options_to_exclude)
}
\arguments{
\item{data}{data table of class \strong{be_monk}}

\item{method}{"q20" if the 0.2 shall be compared; "median" if the median shall be compared.}

\item{answer_options_to_exclude}{a vector of answer options that shall not be included in the analysis}
}
\value{
data table with 0.2-quantiles/medians in times for each question-interviewer-combination
}
\description{
This function investigates if an interviewer has a very low 0.2-quantile or median in any question.
}
\details{
If an interviewer has many low medians, he/she could be a complete falsifier. If an interviewer has many low 0.2-quantiles, but not low medians, he/she could be a partly falifier.
}
