% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_speeders_q.R
\name{inv_speeders_q}
\alias{inv_speeders_q}
\title{Test for fastest answers}
\usage{
inv_speeders_q(data, quantil_of_fastest_interviews = 0.05,
  quantil_of_fastest_interviewer = 0.9, min_of_fast_interviews = 5,
  answer_options_to_exclude)
}
\arguments{
\item{data}{data table of class \strong{be_monk}}

\item{quantil_of_fastest_interviews}{quantil of fastest time stamps for a specific question within all interviewers}

\item{quantil_of_fastest_interviewer}{quantil of presence of specific interviewer with the fastest interval}

\item{min_of_fast_interviews}{min of fast answers within the fastest quantil of all interviews that an invertviewer is regarded conspicious for this question}

\item{quantil_of_fastest_interviewer}{quantil of presence of specific interviewer with the fastest interval}
}
\value{
list containing quantil_speeders_dt (data.table with binary variable if interviewer was conspicious regarding the fastest interviews for each question-interviewer-combination) & dat_long (detailed information of fast interviews)
}
\description{
This function tests if an interviewer has provided conspicious many fast answer times within the fastest answer times of all interviews conducted by all interviewers.
}
\details{
An interviewer conspicious for a specific question has produced many extreme results for this specific question. If this interviewer is also conspicious according to method "q20" in \code{\link{inv_speeders_median}}, this interviewer could partly falsify this question. On the other hand if an interviewer is not additionally conspicious due to method "q20", he or she has produced extreme results, but the majority of interviews was not salient. Extreme results could be therefore outcomes of other reasons, e.g. clicking through the questionnaire sessions to get acquainted with the questions.
}
