% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_statsurv.R
\name{read_in_statsurv}
\alias{read_in_statsurv}
\title{In-house read-in Function for Statistics Austria}
\usage{
read_in_statsurv(file, condition, mode_variable, edge_items, key_variables,
  INT = NULL)
}
\arguments{
\item{file}{path to survey/time data}

\item{condition}{condition of interviews to be analysed}

\item{mode_variable}{name of the variable indicating the mode}

\item{edge_items}{character vector of two items: first and last question to be analysed}

\item{key_variables}{a character vector of variables of high importance e.g. gate_questions}
}
\value{
object of class be_monk
}
\description{
This function is only relevant for users within Statisticis Austria
}
