% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_function.r
\name{sum_function}
\alias{sum_function}
\title{Summarize Function for RMD -Files}
\usage{
sum_function(dat, method)
}
\arguments{
\item{dat}{currently data.table but in construction for classes}

\item{method}{currently method of inv function, but will be skipped as soon as classes are introduced}
}
\value{
in construction: list of two data.tables (one with detailed information and one not)
}
\description{
This function summarizes results of the inv-functions for the Rmd:Report-Files
}
