% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_data_test}
\alias{time_data_test}
\title{Dummy Time Data}
\format{A data table with 4525 Interviews of 141 Variables.
\describe{
each variable is a dummy survey variable
}}
\usage{
time_data_test
}
\description{
Created dummy Time data of 4525 Observations of 141 Variables.
Each observation is an interview with a different person/household/enterprise/... . Values are times needed on specific survey questions.
}
\keyword{datasets}
